nassh: add an ArrayBuffer sink
We have a number of scenarios where we accept fragments of data (as
ArrayBuffers) that we need to merge into a single ArrayBuffer before
we can use it.  Define a new buffer type to transparently support
that use case.

This will replace a lot of code that is doing string concatenation
today as we migrate things from base64 strings to ArrayBuffers.

This version maintains an internal buffer of data that is copied to
during the write phase.  The read & ack phases just update the typed
array views to that buffer.  Another implementation would accept the
arraybuffer pointers during write and then coalesce the data into a
new buffer during the read step.  Performance measurements showed it
to be slower.

Change-Id: I06f1b88cea9e356e49aa0c63c0be31661845dd11