nassh: start a generic worker pipeline helper
We have a number of places where we want to read from one source and
write the data somewhere else.  We usually handle this by reading then
writing in a serial loop.  This can be a bit slow and intermittent
latencies bog everything down.  We also have code that will batch all
the possible reads, resolve them in parallel, then batch up the writes
and resolve them directly.  This is slightly better, but the code also
has bugs where it assumes all the reads resolve in the same order even
if the JS API has no guarantee.

Thus the main processing logic has been ad-hoc to keep it simple, and
because the read/write combos have varied.  Sometimes we read from the
network, sometimes from packets, sometimes from files, and conversely
we write to a variety of places.

Lets create a worker class just for pipelining the main logic.  This
allows users to focus on the producer & consumer components and leave
the rest of the higher logic to this single class.

Change-Id: Id9532710719f535eca9656a0c0fadaa2c6c3f4ef